+++ internal/releaseversion/spec.go
@@ -1,0 +1,11 @@
+package releaseversion
+
+import (
+	"github.com/blang/semver"
+	"github.com/giantswarm/apiextensions/v2/pkg/apis/release/v1alpha1"
+)
+
+type release struct {
+	Version *semver.Version
+	CR      *v1alpha1.Release
+}